refactor: 베럴패턴적용
