+++ src/entities/schedule/models/detailed-schedule.API.ts
@@ -1,6 +1,6 @@
 import axiosInstance from '@/shared/api/axios-instance'
 
-import type { Schedule } from '@/entities/schedule'
+import type { Schedule } from '@/entities/schedule/models'
 
 export async function getDetailedSchedule(scheduleId: string): Promise<Schedule> {
   const res = await axiosInstance(`/schedules/${scheduleId}`)
