+++ src/entities/schedule/models/index.ts
@@ -1,0 +1,4 @@
+export * from './schedule.types'
+export * from './get-group-by-date'
+export * from './detailed-schedule.API'
+export * from './schedule.query'
