+++ src/features/my-schedule/hooks/use-date-schedules.ts
@@ -1,7 +1,6 @@
 import { useSuspenseQuery } from '@tanstack/react-query'
 
-import { groupByDate } from '@/entities/schedule'
-import { scheduleQueryKeys } from '@/entities/schedule/models/schedule.query'
+import { groupByDate, scheduleQueryKeys } from '@/entities/schedule/models'
 import { useUserStore } from '@/entities/user/models/use-user-store'
 
 import { getMySchedule } from '../api'
